+++ src/main/java/org/springframework/social/partnercenter/PartnerCenterAdmin.java
@@ -7,6 +7,7 @@ import org.springframework.social.partnercenter.api.customer.user.role.Directory
 import org.springframework.social.partnercenter.api.customer.user.role.RoleOperations;
 import org.springframework.social.partnercenter.api.order.AdminOrderOperations;
 import org.springframework.social.partnercenter.api.order.subscription.AdminSubscriptionOperations;
+import org.springframework.social.partnercenter.api.relationships.AdminRelationshipOperations;
 import org.springframework.social.partnercenter.api.support.SupportOperations;
 
 public interface PartnerCenterAdmin extends PartnerCenter {
@@ -20,4 +21,6 @@ public interface PartnerCenterAdmin extends PartnerCenter {
 	SupportOperations getSupportOperations();
 	DirectoryRoleOperations getDirectoryRoleOperations();
 	RoleOperations getRoleOperations();
+
+	AdminRelationshipOperations getRelationshipOperations();
 }